chore: mainpage_파일구조 및 schema 변경에 따른 코드 수정
