+++ src/pages/MainPage.jsx
@@ -2,7 +2,7 @@
 import React from 'react'
 import Header from './main/Header'
 import Footer from './main/Footer'
-import HomeBody from './main/HomeBody'
+import HomeBody from './main/HomeBody/HomeBody'
 
 function MainPage() {
   return (
