+++ src/pages/main/Header.jsx
@@ -1,10 +1,10 @@
 /* eslint-disable import/no-unresolved */
 import React from 'react'
 import { Box } from '@mui/material'
-import HeaderLogo from './HeaderLogo'
-import HeaderSearchBox from './HeaderSearchBox'
-import FunctionOptions from './FunctionOptions'
-import NavigationBar from './NavigationBar'
+import HeaderLogo from './Header/HeaderLogo'
+import HeaderSearchBox from './Header/HeaderSearchBox'
+import FunctionOptions from './Header/FunctionOptions'
+import NavigationBar from './Header/NavigationBar'
 import Filtering from './Filtering/Filtering'
 
 function Header() {
