+++ src/pages/main/HomeBody/HotLodging.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/extensions */
 /* eslint-disable react/prop-types */
 /* eslint-disable react/no-array-index-key */
 import { useState, useEffect } from 'react'
@@ -7,7 +8,7 @@ import KingBedIcon from '@mui/icons-material/KingBed'
 import BookOnlineIcon from '@mui/icons-material/BookOnline'
 import HotTab from './HotTab'
 import HotLodgingList from './HotLodgingList'
-import { fetchData } from '../../fetch/search'
+import { fetchData } from '../../../fetch/search.js'
 
 function HotLodging() {
   const [selectedLocale, setSelectedLocale] = useState('서울')
